add support for silencing deregistration events
