+++ backend/keepalived/deregisterer_test.go
@@ -5,9 +5,12 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/sensu/sensu-go/backend/store"
+	"github.com/sensu/sensu-go/backend/store/cache"
 
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/testing/mockbus"
+	"github.com/sensu/sensu-go/testing/mockcache"
 	"github.com/sensu/sensu-go/testing/mockstore"
 	"github.com/sensu/sensu-go/types"
 	"github.com/stretchr/testify/assert"
@@ -43,13 +46,15 @@ func TestDeregister(t *testing.T) {
 func TestDeregistrationHandler(t *testing.T) {
 	assert := assert.New(t)
 
+	mockCache := &mockcache.MockCache{}
 	mockStore := &mockstore.MockStore{}
 	mockBus := &mockbus.MockBus{}
 
 	adapter := &Deregistration{
-		EventStore:  mockStore,
-		EntityStore: mockStore,
-		MessageBus:  mockBus,
+		EventStore:    mockStore,
+		EntityStore:   mockStore,
+		MessageBus:    mockBus,
+		SilencedCache: mockCache,
 	}
 
 	entity := types.FixtureEntity("entity")
@@ -58,6 +63,12 @@ func TestDeregistrationHandler(t *testing.T) {
 		Handler: "deregistration",
 	}
 	check := types.FixtureCheck("check")
+
+	mockCache.On("Get", "default").Once().Return(
+		[]cache.Value{
+			{Resource: corev2.FixtureSilenced("*:deregistration")},
+		},
+	)
 
 	mockStore.On("GetEventsByEntity", mock.Anything, entity.Name, &store.SelectionPredicate{}).Return([]*types.Event{}, nil)
 	mockStore.On("DeleteEventByEntityCheck", mock.Anything, entity.Name, check.Name).Return(nil)
@@ -68,6 +79,7 @@ func TestDeregistrationHandler(t *testing.T) {
 		assert.Equal("deregistration", event.Entity.Deregistration.Handler)
 		assert.Equal("deregistration", event.Check.Name)
 		assert.Equal(0, len(event.Check.Subscriptions))
+		assert.Equal(true, event.IsSilenced(), "event is not silenced")
 		if event.Timestamp == 0 {
 			t.Fatal("event timestamp is nil, expected a timestamp in the deregistration event")
 		}
